requirements.txt
openai>=1.14.0
pillow>=10.0.0
lancedb>=0.4.0,<0.5.0
langgraph>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.2.0
jsonlines>=3.1.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
PyYAML>=6.0
aiofiles>=23.0.0
httpx>=0.24.0
pyarrow>=12.0.0 